+++ src/components/CommitLintComponent.ts
@@ -1,6 +1,7 @@
 import { TextFile } from 'projen';
 import { TypeScriptProject } from 'projen/lib/typescript';
-import { IProjectComponent, Scripts } from '../types';
+import { IProjectComponent } from '../types/component';
+import { Scripts } from '../types/script';
 
 /**
  * Configures the CommitLint templates, settings and npm scripts for the project.
