--- src/types/general.ts
@@ -1,5 +1,0 @@
-export type LintStagedConfig = {
-  'lint-staged': {
-    [pattern: string]: string[];
-  };
-};
